# api.yml
openapi: "3.0.0"
info:
  description: Conditions API expose CRUD actions to condition objects on servers
  title: Condition orchestrator API
  version: 1.0.0
paths:
  /servers/{id}/conditions:
    get:
      summary: Returns all conditions set on a server
      operationId: serverConditionList
      parameters:
        - name: id
          in: path
          description: Server ID
          required: true
          schema:
            $ref: "#/components/schemas/ServerID"
      responses:
        "200":
          description: server conditions response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
  /server/foo:
    post:
      summary: Sets a condition on a server
      operationId: serverConditionSet
      parameters:
        - name: id
          in: path
          description: Server ID
          required: true
          schema:
            $ref: "#/components/schemas/ServerID"
      requestBody:
        description: Condition to add to server
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewCondition"
      responses:
        "200":
          description: Set condition on server response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Condition"
components:
  schemas:
    ServerID:
      type: string
      format: uuid
    ConditionState:
      type: string
      enum:
        - pending
        - active
        - failed
        - succeeded
    ConditionKind:
      type: string
      enum:
        - firmwareInstall
        - inventoryOutofband
    Condition:
      type: object
      properties:
        Kind:
          type: string
          format: ConditionKind
          oneOf:
            - $ref: "#/components/schemas/ConditionKind"
          readOnly: true
        Parameters:
          type: object
          additionalProperties: {}
        State:
          type: string
          oneOf:
            - $ref: "#/components/schemas/ConditionState"
        Status:
          type: object
          additionalProperties: {}
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
    NewCondition:
      required:
        - kind
        - parameters
      properties:
        kind:
          type: string
          format: ConditionKind
          oneOf:
            - $ref: "#/components/schemas/ConditionKind"
        parameters:
          type: object
          additionalProperties: {}
    Server:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        Conditions:
          type: array
          items:
            $ref: "#/components/schemas/Condition"